requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
finvizfinance>=0.14.0
vaderSentiment>=3.3.2
polygon-api-client>=1.12.0
yfinance>=0.2.0
scipy>=1.11.0
py-vollib>=1.0.0
